Remove context
